# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel API",
        "description": "API para gestionar hoteles, habitaciones y reservas. Protegida por API Key (X-API-KEY).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/hotels/{hotelId}/rooms": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Listar las habitaciones de un determinado hotel",
                "description": "Devuelve el listado de habitaciones pertenecientes al hotel indicado",
                "operationId": "442cf7acae7456f95967c56db11459c4",
                "parameters": [
                    {
                        "name": "hotelId",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de habitaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Habitación Matrimonial"
                                            },
                                            "capacity": {
                                                "type": "integer",
                                                "example": 2,
                                                "nullable": true
                                            },
                                            "room_type": {
                                                "type": "string",
                                                "example": "standard",
                                                "nullable": true
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 120.5,
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Crear una habitación de un hotel",
                "description": "Crea una nueva habitación asociada al hotel indicado",
                "operationId": "8bfc7a4e1654d05a797854e18881532f",
                "parameters": [
                    {
                        "name": "hotelId",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Habitación Matrimonial"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "room_type": {
                                        "type": "string",
                                        "example": "suite",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Habitación creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Habitación Matrimonial"
                                        },
                                        "capacity": {
                                            "type": "integer",
                                            "example": 2,
                                            "nullable": true
                                        },
                                        "room_type": {
                                            "type": "string",
                                            "example": "suite",
                                            "nullable": true
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 150,
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "property": null
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}